Moving topic provider up to common
