+++ roslibrust_common/Cargo.toml
@@ -4,11 +4,15 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+# Used for error handling
 thiserror = "2.0"
 anyhow = "1.0"
+# Used as basis for serialization
 serde = { version = "1.0", features = ["derive"] }
 # Used for md5sum calculation
 md5 = "0.7"
+# Used for async trait definitions
+futures = "0.3"
 
 # THESE DEPENDENCIES WILL BE REMOVED
 # We're currently leaking these error types in the "generic error type"
