+++ roslibrust_zenoh/Cargo.toml
@@ -7,11 +7,7 @@ edition = "2021"
 
 [dependencies]
 tokio = { version = "1.41", features = ["rt-multi-thread", "sync", "macros"] }
-# We currently need to depend on these crates for trait information
-# I think we should move topic_provider and the generic message traits into their own crate
-# Then more things can work with them generically with less compile time dependencies
-roslibrust = { path = "../roslibrust", features = ["topic_provider"] }
-roslibrust_codegen = { path = "../roslibrust_codegen" }
+roslibrust_common = { path = "../roslibrust_common" }
 zenoh = "1.0"
 hex = "0.4"
 anyhow = "1.0"
@@ -20,7 +16,9 @@ log = "0.4"
 
 [dev-dependencies]
 env_logger = "0.11"
+# Used to access traits for examples
+roslibrust = { path = "../roslibrust" }
 # Used to generate message types for the examples
 roslibrust_codegen_macro = { path = "../roslibrust_codegen_macro" }
 # Relied on by generate types in the macro this should be cleaned up
-roslibrust_common = { path = "../roslibrust_common" }
+roslibrust_codegen = { path = "../roslibrust_codegen" }
